ENH: Moved hx and vmd drafts to output
